% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_prepare.R
\name{create_InformativeGenes}
\alias{create_InformativeGenes}
\title{Create an InformativeGenes object}
\usage{
create_InformativeGenes(tumor_ct, tumor_lab, interested_subtype)
}
\arguments{
\item{tumor_ct}{a data frame or matrix of the un-normalized (estimated) counts of sequencing reads or fragments from the tumor data with
Rows for genes with row names of HGNC gene symbols and columns for samples.}

\item{tumor_lab}{a vector of strings indicating the labels of the tumor samples.}

\item{interested_subtype}{a string indicating the interested tumor subtype, and it should be included in the tumor labels.}
}
\value{
An \code{InformativeGenes} object.
}
\description{
This function runs differential expression analysis (DESeq2) and gene set enrichment analysis (GSEA) to generate a InformativeGenes object,
which is necessary for creating the GULL object.
}
\references{
Love, Michael I., Wolfgang Huber, and Simon Anders. "Moderated estimation of fold change and dispersion for RNA-seq data with DESeq2." Genome biology 15, no. 12 (2014): 1-21.
}
