% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_pathway_analysis.R
\name{pathway_analysis}
\alias{pathway_analysis}
\title{Pathway specific selection analysis}
\usage{
pathway_analysis(object)
}
\arguments{
\item{object}{A \code{GULL} object.}
}
\value{

}
\description{
Pathway specific selection analysis
}
